Add runStream test, move reduce/observe/drain to test/helper
