+++ packages/core/test/combinator/mergeConcurrently-test.js
@@ -5,7 +5,7 @@ import { mergeMapConcurrently, mergeConcurrently } from '../../src/combinator/me
 import { periodic } from '../../src/source/periodic'
 import { take } from '../../src/combinator/slice'
 import { constant } from '../../src/combinator/transform'
-import { drain } from '../../src/combinator/observe'
+import { drain } from '../helper/observe'
 import { just as just } from '../../src/source/core'
 import { fromArray } from '../../src/source/fromArray'
 import { collectEventsFor } from '../helper/testEnv'
